OK.. in rig control object I have written a QRG validation routine that attempts to resolve any string *I* can think of that may be passed to it into a QRG integer value in Hertz.  It has two modes, strict and lax.  In strict mode the QRG will not validate if it is not in an amateur band from 160M to 33cm excluding 60M.  In lax mode it just needs to resolve to something 'sane'.
In the main GUI when in CAT (rig control) mode an invalid qrg will lead to the QRG readout box flashing between the 'invalid' QRG and 0.0

This was actually a happy bug as it was unintended, but, is being left as is for a visual indication that the QRG value is not valid for the purposes of JT65 operation.  60M is a bit of a quandary in that I consider any QRG in its range to be invalid and exclude spots on that band to the RB system.  I know some regions have fairly liberal privileges on the band, but, I don't want to unwittingly lead some 'misinformed' souls in regions such as the USA to break the rules and fire up JT65 on 60M.  So, for now I am limiting JT65-HF operation to the 160,80,40,30,20,17,15,12,10,6,2,1,25 meter bands plus 70 and 33 cm bands.  This, of course, only applies to the spotting networks.  You can use JT65-HF on any QRG but you can only send spots for those bands listed above.

+++ win32/GUI-fpc-lazarus/guidedconfig.pas
@@ -954,7 +954,7 @@ end;
 
 procedure TForm7.buttonContinue6Click(Sender : TObject);
 var
-   tstflt : Double;
+   tstflt : Single;
 begin
      // Called from Macros Tab
      // Now... can I learn to iterate the controls via code or do I have to do this by 'hand'
@@ -2158,7 +2158,7 @@ end;
 procedure TForm7.readConfig(cfgname : string);
 var
    i       : Integer;
-   tmpflt  : double;
+   tmpflt  : single;
    foo     : String;
 begin
      tmpflt := 0.0;
